Bubble Sort
Bubble sorting is to move the small elements forward or the transfer of
large elements back. Comparison is the comparison of two adjacent
elements, exchange also occurs between the two elements. So, if the two
elements are equal, I think you will not bother to exchange the two of
them both; if two equal elements are not adjacent, then even through the
front of the two exchange to two adjacent, this Time will not be
exchanged, so the same elements before and after the order has not
changed, so the bubble sort is a stable sorting algorithm.